Add Pause After Each Page option
When enabled, print jobs will be suspended after every page. Press the paper button to continue printing. This may be helpful for manual two-sided printing on personal printers without duplexing support.

On this commit the printer will pause after the last page, press the paper button to finish the job.